requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
opencv-python
torch>=1.12.0
diffusers[torch]==0.20.0
transformers>=4.30.0
accelerate>=0.18.0
safetensors
scikit-learn
